proyecto/requirements.txt
asgiref==3.8.1
Django==5.1.3
mysql-connector-python==9.2.0
packaging==25.0
Pillow==11.1.0
python-decouple==3.8
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.13.2
gunicorn==23.0.0
whitenoise==6.9.0
Brotli==1.1.0
uvicorn==0.34.3